requirements.txt
fastapi
uvicorn
sqlmodel
asyncpg
jinja2
httpx
python-multipart
python-dotenv
aiosqlite
pydantic[email]
markdown
passlib[bcrypt]
slowapi
pydantic-settings
async-lru
google-genai
mcstatus
python-a2s
itsdangerous
cachetools
aiofiles